% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_prepare.R
\name{assoc_prepare}
\alias{assoc_prepare}
\title{Generate frequency table for association measure}
\usage{
assoc_prepare(colloc_out = NULL, stopword_list = NULL)
}
\arguments{
\item{colloc_out}{The output list of \code{\link{colloc_leipzig}}.}

\item{stopword_list}{Character vectors containing list of stopwords to be removed from the collocation measures.}
}
\value{
A tbl_df of two columns. One of them is nested columns with input-data for row-wise Fisher-Exact Test with \code{\link{collex_fye}}.
}
\description{
The function to produce frequency table required as input for association measures for collocations
}
\examples{
\dontrun{
 assoc_tb <- assoc_prepare(colloc_leipzig_output, stopword_list = NULL)
}

}
