% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assoc_prepare_dca.R
\name{assoc_prepare_dca}
\alias{assoc_prepare_dca}
\title{DCA input table}
\usage{
assoc_prepare_dca(assoc_tb_df = NULL)
}
\arguments{
\item{assoc_tb_df}{The nested tibble output from \code{\link{assoc_prepare}}}
}
\value{
A tibble with columns for collocates, frequency of the collocates with Construction/Word A and B, other frequencies in the cross-tabulation (i.e. Cell C and D), and the Expected frequencies of the collocates with Construction/Word A as the reference item.
}
\description{
A function to generate input table for performing Distinctive Collexeme Analysis.
}
\examples{
\dontrun{
# run assoc_prepare() first from the output of colloc_leipzig()
 assoc_tb <- assoc_prepare(colloc_out = colloc_leipzig_output,
                           window_span = "r1",
                           per_corpus = FALSE,
                           stopword_list = NULL,
                           float_digits = 3L)

# then run assoc_prepare_dca()
dca_tb <- assoc_prepare_dca(assoc_tb)
}
}
