% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_file_input}
\alias{check_file_input}
\title{Verify the input file}
\usage{
check_file_input(path_input = NULL, list_input = NULL)
}
\arguments{
\item{path_input}{Receive input from the \code{leipzig_path} argument of \code{colloc_leipzig}.}

\item{list_input}{Receive input from the \code{leipzig_corpus_list} argument of \code{colloc_leipzig}.}
}
\description{
Utility function to check whether \code{colloc_leipzig} receive a file path input or a list-corpus input.
    It is called internally by \code{colloc_leipzig}.
}
