% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{check_type_input}
\alias{check_type_input}
\title{Check input type}
\usage{
check_type_input(input = NULL, input_names = NULL, indices = NULL)
}
\arguments{
\item{input}{Either a character vector of Leipzig Corpora file path or a list whose elements correspond to the Leipzig Corpora File loaded into R.}

\item{input_names}{Names of the corpus file}

\item{indices}{For-loop index within \code{colloc_leipzig}}
}
\value{
A character vector of corpus sentences.
}
\description{
Utility function to check type of object for the input of \code{colloc_leipzig}.
    It is called internally by \code{colloc_leipzig}.
}
