% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq.R
\name{collex_chisq}
\alias{collex_chisq}
\title{Chi-square test for association measure}
\usage{
collex_chisq(df, collstr_digit = 3)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare}}.}

\item{collstr_digit}{The floating digits of the collostruction strength. The default is \code{3}.}
}
\value{
A tibble consisting of the collocates (column \code{w}),
    co-occurrence frequencies with the node (column \code{a}),
    the expected co-occurrence frequencies with the node (column \code{a_exp}),
    the direction of the association (e.g., \emph{attraction} or \emph{repulsion}) (column \code{assoc}),
    and the chi-square-based collostruction strength (column \code{chisq}).
}
\description{
Compute collostructional strength based on the chi-square contribution.
    It internally calls the utility function \code{\link{chisq_compute}} and performs a row-wise computation using \code{\link[purrr]{map_dbl}}.
}
\examples{
out <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
                      pattern = "mengatakan",
                      window = "r",
                      span = 3L,
                      save_interim = FALSE)
assoc_tb <- assoc_prepare(colloc_out = out, stopword_list = stopwords)
am_chisq <- collex_chisq(df = assoc_tb, collstr_digit = 3)
}
