% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fye.R
\name{collex_fye}
\alias{collex_fye}
\title{Fisher's Exact Test}
\usage{
collex_fye(df, collstr_digit = 3)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare}}.}

\item{collstr_digit}{The numeric vector for floating digits of the collostruction strength. The default is \code{3}.}
}
\value{
A tibble consisting of the collocates (column \code{w}),
    co-occurrence frequencies with the node (column \code{a}),
    the expected co-occurrence frequencies with the node (column \code{a_exp}),
    the direction of the association (e.g., \emph{attraction} or \emph{repulsion}) (column \code{assoc}),
    the collostruction strength (column \code{collstr}),
    and two uni-directional association measures of \emph{Delta P}.
}
\description{
Perform one-tailed Fisher's Exact test for the collostruction/collocation strength.
    It internally calls the utility function \code{\link{fye_compute}} and performs a row-wise computation using \code{\link[purrr]{map_dbl}}.
    The \emph{p}-value is log-transformed to the base of ten as in the \emph{Collostructional Analysis}.
}
\examples{
out <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
                      pattern = "mengatakan",
                      window = "r",
                      span = 3L,
                      save_interim = FALSE)
assoc_tb <- assoc_prepare(colloc_out = out, stopword_list = stopwords)
am_fye <- collex_fye(df = assoc_tb, collstr_digit = 3)

}
