% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fye_dca.R
\name{collex_fye_dca}
\alias{collex_fye_dca}
\title{Fisher's Exact Test for DCA}
\usage{
collex_fye_dca(df = NULL, collstr_digit = 3L)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare_dca}}.}

\item{collstr_digit}{The integer for floating digits of the collostruction strength. The default is \code{3L}.}
}
\value{
A data frame containing the collocates, their frequency with Construction/Word A and B, p-value of the Fisher Test, Collostruction Strength, and information on the distinctiveness of the collocates with the constructions.
}
\description{
Perform one-tailed Fisher's Exact test for the \emph{Distinctive Collexeme/Collocates Analysis}.
}
\examples{
\dontrun{

# get the input format via assoc_prepare_dca()
dca_tb <- assoc_prepare_dca(assoc_tb)

# perform the DCA
dca_res <- collex_fye_dca(dca_table)

}
}
