% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collex_llr.R
\name{collex_llr}
\alias{collex_llr}
\title{Log-likelihood ratio}
\usage{
collex_llr(df = NULL, collstr_digit = 3)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare}}.}

\item{collstr_digit}{The numeric vector for floating digits of the collostruction strength. The default is \code{3}.}
}
\value{
A tibble consisting of the collocates (column \code{w}),
    co-occurrence frequencies with the node (column \code{a}),
    the expected co-occurrence frequencies with the node (column \code{a_exp}),
    the direction of the association (e.g., \emph{attraction} or \emph{repulsion}) (column \code{assoc}),
    the log-likelihood ratio as the association/collostruction strength measure (column \code{llr}).
    and two uni-directional association measures of \emph{Delta P}.
}
\description{
Perform log-likelihood ratio as the association/collostruction strength.
}
\examples{
out <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
                      pattern = "mengatakan",
                      window = "r",
                      span = 3L,
                      save_interim = FALSE)
assoc_tb <- assoc_prepare(colloc_out = out, stopword_list = stopwords)
am_llr <- collex_llr(df = assoc_tb, collstr_digit = 3)
}
