% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collex_MI.R
\name{collex_MI}
\alias{collex_MI}
\title{Mutual Information score}
\usage{
collex_MI(df, collstr_digit = 3)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare}}.}

\item{collstr_digit}{The numeric vector for floating digits of the collostruction strength. The default is \code{3}.}
}
\value{
A tibble consisting of the collocates (column \code{w}),
    co-occurrence frequencies with the node (column \code{a}),
    the expected co-occurrence frequencies with the node (column \code{a_exp}),
    the direction of the association (e.g., \emph{attraction} or \emph{repulsion}) (column \code{assoc}),
    the Mutual Information score (column \code{MI}),
    and two uni-directional association measures of \emph{Delta P}.
}
\description{
The function to compute collocation association measure with Mutual Information.
}
\examples{
out <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
    pattern = "ke", # it is a preposition meaning 'to(wards)'
    window = "r",
    span = 2L,
    save_interim = FALSE)
assoc_tb <- assoc_prepare(colloc_out = out,
     stopword_list = collogetr::stopwords[collogetr::stopwords != "ke"])

collex_MI(assoc_tb)


}
