% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collex_TScore.R
\name{collex_TScore}
\alias{collex_TScore}
\title{T-Score association measure}
\usage{
collex_TScore(df)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare}}.}
}
\value{
A tibble consisting of the collocates (column \code{w}),
    co-occurrence frequencies with the node (column \code{a}),
    the expected co-occurrence frequencies with the node (column \code{a_exp}),
    the direction of the association (e.g., \emph{attraction} or \emph{repulsion}) (column \code{assoc}),
    the T-Score (column \code{tscore}),
    and two uni-directional association measures of \emph{Delta P}.
}
\description{
The function to compute collocation association measure with T-Score.
}
\examples{
out <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
    pattern = "ke", # it is a preposition meaning 'to(wards)'
    window = "r",
    span = 2L,
    save_interim = FALSE)
assoc_tb <- assoc_prepare(colloc_out = out,
     stopword_list = collogetr::stopwords[collogetr::stopwords != "ke"])

collex_TScore(assoc_tb)

}
