% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colloc_leipzig.R
\name{colloc_leipzig}
\alias{colloc_leipzig}
\title{Generate window-span collocates for the Leipzig Corpora}
\usage{
colloc_leipzig(
  leipzig_path = NULL,
  leipzig_corpus_list = NULL,
  pattern = NULL,
  case_insensitive = TRUE,
  window = "b",
  span = 2,
  split_corpus_pattern = "([^a-zA-Z-¬]+|--)",
  to_lower_colloc = TRUE,
  save_interim = FALSE,
  freqlist_output_file = "collogetr_out_1_freqlist.txt",
  colloc_output_file = "collogetr_out_2_collocates.txt",
  corpussize_output_file = "collogetr_out_3_corpus_size.txt",
  search_pattern_output_file = "collogetr_out_4_search_pattern.txt"
)
}
\arguments{
\item{leipzig_path}{Character strings of (i) file names of the Leipzig corpus if they are in the working directory, or (ii) the complete file path to each of the Leipzig corpus files.}

\item{leipzig_corpus_list}{Specify this argument if each Leipzig corpus file has been loaded as R object and acts as an element of a \bold{named list}.
Example of this type of data-input can be seen in \code{data("demo_corpus_leipzig")}.
So specify either \code{leipzig_path} OR \code{leipzig_corpus_list} and set one of them to \code{NULL}.}

\item{pattern}{Character vector input containing a set of exact word forms.}

\item{case_insensitive}{Logical; whether the search for the \code{pattern} ignores case (\code{TRUE} -- default) or not (\code{FALSE}).}

\item{window}{Character; window-span direction of the collocates: \code{"r"} ('\bold{right} of the node'), \code{"l"} ('\bold{left} of the node'), or the default is \code{"b"} ('both \bold{left} and \bold{right} context-window').}

\item{span}{A numeric vector indicating the span of the collocate scope. The default is \code{2} words around the node word.}

\item{split_corpus_pattern}{Regular expressions used to tokenise the corpus into word-vector.
The default regex is \code{"([^a-zA-Z-\u00AC]+|--)"}. The character \code{"\u00AC"} is a hexademical version of \code{"¬"}, which may occur in the Leipzig Corpora as separator between root and suffixes of a word, in addition to hypen.
This procedure supports the vectorised method of the function to generate the collocate of the search pattern.}

\item{to_lower_colloc}{Logical; whether to lowercase the retrieved collocates and the nodes (\code{TRUE} -- default) or not (\code{FALSE}).}

\item{save_interim}{Logical; whether to save interim results into plain text files or not (\code{FALSE} -- default).}

\item{freqlist_output_file}{Character strings for the name of the file for the word frequency in a corpus.}

\item{colloc_output_file}{Character strings for the name of the file for the raw collocate table.}

\item{corpussize_output_file}{Character strings for the name of the file for the total word-size of a corpus.}

\item{search_pattern_output_file}{Character strings for the name of the file for the search_pattern.}
}
\value{
List of raw collocate items, frequency list of all words in the loaded corpus files, the total word tokens in each loaded corpus, and the search pattern.
}
\description{
The function produces tibble-output collocates for the Leipzig corpus files.
}
\examples{
collout <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
                     pattern = "mengatakan",
                     window = "r",
                     span = 3,
                     save_interim = FALSE)
# collout <- colloc_leipzig(leipzig_corpus_path = c('path_to_corpus1.txt',
#                                                     'path_to_corpus2.txt'),
#                             pattern = "mengatakan",
#                             window = "r",
#                             span = 3,
#                             save_interim = TRUE # save interim output file
#                             # you need to specify path in the argument
#                             # with \code{...output_file}
#                             )


}
