% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coll_sentmatch.R
\name{colloc_sentmatch}
\alias{colloc_sentmatch}
\title{Sentence match retriever}
\usage{
colloc_sentmatch(collout, colloc = NULL, wspan = NULL,
  nodeword = NULL, sampled = NULL)
}
\arguments{
\item{collout}{List output of \code{\link{colloc_leipzig}}.}

\item{colloc}{Character vector of the collocate(s) whose sentence match(es) to be retrieved.}

\item{wspan}{Character vector of the window span in which the collocates occur. Default to \code{NULL}, which will retrieve the collocate's occurrence in all span.}

\item{nodeword}{Character vector specifying one of the nodewords if search parameter in \code{\link{colloc_leipzig}} includes more than one nodeword.}

\item{sampled}{Integer vector indicating the number of random sample of the sentence match to be retrieve. Default to \code{NULL}, which will retrieve all sentence-matches.}
}
\value{
Character vector of sentence-match(es).
}
\description{
The function extract full sentence-match for a (set of significant) collocate(s) for a given nodeword. Given that
}
\examples{
collout <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
                     pattern = "mengatakan",
                     window = "r",
                     span = 3,
                     save_interim = FALSE)

colloc_sentmatch(collout,
                 colloc = "bahwa",
                 sampled = 10)

# This will produce message indicating that
# the queried sample number is higher than
# the sentence match for "akan"
colloc_sentmatch(collout,
                 colloc = "akan",
                 sampled = 10)
}
