% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coll_sentmatch_tagged.R
\name{colloc_sentmatch_tagged}
\alias{colloc_sentmatch_tagged}
\title{Sentence match retriever with tagged collocations}
\usage{
colloc_sentmatch_tagged(collout, colloc = NULL)
}
\arguments{
\item{collout}{List output of \code{\link{colloc_leipzig}}.}

\item{colloc}{Character vector of the collocate(s) whose sentence match(es) to be retrieved.}
}
\value{
A data frame with the following variables/columns:
\itemize{
  \item \code{corpus_names} - corpus file name
  \item \code{sent_id} - sentence number of the collocate matches
  \item \code{w} - the collocate whose full sentence match with the node is retrieved via the function
  \item \code{span} - the window-span of the collocate in relation to the node word
  \item \code{node} - the node word
  \item \code{sent_match_tagged} - tagged sentence matches for the collocate of interest with the node
  \item \code{coll_pattern} - extracted collocate-node pattern from the sentence matches
}
}
\description{
This is an expanded function for \code{\link{colloc_sentmatch}} to retrieve sentence matches in which a given collocate is found.
The extended features in \code{\link{colloc_sentmatch_tagged}} (but not available in \code{\link{colloc_sentmatch}}) are:
\enumerate{
  \item a tibble/data frame output format (similar to the \code{colloc_df}, which is one of the outputs of \code{\link{colloc_leipzig}})
  \item tagging for the collocates (with \code{"<c>...</c>" tag}) and the nodes (\code{"<n>...</n>"}) in the sentence matches
  \item a column called \code{"coll_pattern"} containing the collocate-node pattern extracted via regular expression on the basis of the tagging.
}
}
\details{
\Sexpr[results=rd, stage=render]{lifecycle::badge("experimental")}
}
\examples{
# retrieve the collocate of "sudah" 'already'
collout <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig,
                          pattern = "sudah",
                          window = "b",
                          span = 2,
                          save_interim = FALSE)

# retrieve and tag the sentence match for "sudah" with its collocate "ada" 'exist'
df_sentmatch <- colloc_sentmatch_tagged(collout, colloc = "ada")

# check the tagging in the sentence match
df_sentmatch$sent_match_tagged

# check the extracted collocation pattern and the window span
df_sentmatch[, c("span", "coll_pattern")]

}
\seealso{
\code{\link{colloc_sentmatch}} for untagged and character-vector version of the output, \code{\link{colloc_leipzig}} for collocate retrieval.
}
