% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dca_coll}
\alias{dca_coll}
\title{Demo data for Distinctive Collexeme/Collocate Analysis}
\format{An object of class \code{list} of length 4.}
\usage{
dca_coll
}
\description{
The output of \code{colloc_leipzig()} for illustrating \emph{Distinctive Collexeme/Collocate Analysis} on package README page.
}
\keyword{datasets}
