% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_corpus_leipzig}
\alias{demo_corpus_leipzig}
\title{Leipzig corpora demo}
\format{An object of class \code{list} of length 15.}
\usage{
demo_corpus_leipzig
}
\description{
A random subset of 250 sentences from all unannotated Indonesian Leipzig Corpora files.
}
\keyword{datasets}
