% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_p.R
\name{dP_cue_cxn}
\alias{dP_cue_cxn}
\title{Uni-directional measure \emph{Delta P} (Collocates as the cue).}
\usage{
dP_cue_cxn(df, collstr_digit = NULL)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare}}.}

\item{collstr_digit}{The floating digits of the collostruction strength. It is passed on from \code{\link{collex_fye}} and the default is \code{3}.}
}
\value{
A double numeric vector
}
\description{
Perform uni-directional association measure called \emph{Delta P}. This one measures how likely is the construction/node word given the presence of the collocates/collexemes. That is, the collocates/collexemes are the cue while the construction/node word is the outcome.
    The function is internally called when performing \code{\link{collex_fye}}.
}
