% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fye.R
\name{exp_freq}
\alias{exp_freq}
\title{Get the expected frequency}
\usage{
exp_freq(df, collstr_digit)
}
\arguments{
\item{df}{Nested data frame}

\item{collstr_digit}{The numeric vector for the floating digits to keep. It is passed from \code{\link{assoc_prepare}}.}
}
\value{
A double vector
}
\description{
Generate expected frequency for cell \emph{a} based on the output of Chi-square Test.
    It is called internally via \code{\link{assoc_prepare}}.
}
