% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fye.R
\name{fye}
\alias{fye}
\title{Fisher's Exact Test}
\usage{
fye(df, mpfr_precision = 120, collstr_digit = 3)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare}}.}

\item{mpfr_precision}{Integer indicating the maximal precision to be used in \bold{bits}. This is passed to the \code{precBits} argument of \code{\link[Rmpfr]{mpfr}}.}

\item{collstr_digit}{The floating digits of the collostruction strength. The default is \code{3}.}
}
\value{
A tibble consisting of the collocates (column \code{w}),
    co-occurrence frequencies with the node (column \code{a}),
    the expected co-occurrence frequencies with the node (column \code{a_exp}),
    and the collostruction strength (column \code{collstr}).
}
\description{
Perform one-tailed Fisher's Exact test for the collostruction/collocation strength.
    The \emph{p}-value is log-transformed to the base of ten as in the \emph{Collostructional Analysis}.
}
\examples{
out <- colloc_leipzig(leipzig_corpus_list = demo_corpus_leipzig[1:4],
                      pattern = "\\\\bke\\\\b",
                      window = "r",
                      span = 1,
                      save_interim = FALSE)
assoc_tb <- assoc_prepare(colloc_out = out, stopword_list = stopwords)
am_fye <- fye(df = assoc_tb, mpfr_precision = 120, collstr_digit = 3)
am_fye[order(-am_fye$collstr),]

}
