% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fye.R
\name{fye_compute}
\alias{fye_compute}
\title{Fisher's Exact Test}
\usage{
fye_compute(df)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare}}.}
}
\value{
Numeric vectors of p-values
}
\description{
Perform one-tailed Fisher's Exact test for the collostruction/collocation strength.
    The \emph{p}-value is log-transformed to the base of ten as in the \emph{Collostructional Analysis}.
    This is an internal function called via \code{\link{collex_fye}}.
}
