% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leipzig_corpus_path}
\alias{leipzig_corpus_path}
\title{Balinese newspapers texts}
\format{An object of class \code{character} of length 15.}
\usage{
leipzig_corpus_path
}
\description{
A random ten corpus files of Balinese newspapers texts. These corpus files are raw plain texts without any annotation and which have not been split by sentences.
Indonesian short stories
}
\details{
A random ten corpus files of Indonesian short stories retrieved from online blog. These corpus files are raw plain texts without any annotation and which have not been split by sentences.
Filepath of the Indonesian Leipzig Corpora

Package author's filepath to the plain-texts Indonesian Leipzig Corpora
}
\keyword{datasets}
