% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collex_logOR.R
\name{logOR_compute}
\alias{logOR_compute}
\title{Compute Log10 Odds Ratio}
\usage{
logOR_compute(a, b, c, d)
}
\arguments{
\item{a}{co-occurrence frequency between a given collocate and the nodeword.}

\item{b}{co-occurrence frequency between the collocate and other words.}

\item{c}{co-occurrence frequency between the nodeword and the other words.}

\item{d}{co-occurrence frequency between all other words.}
}
\value{
A double-numeric value of Log10 Odds Ratio score
}
\description{
internal function called by \code{\link{collex_logOR}} to compute Log10 Odds Ratio (Gries 2014).
}
\examples{
# Data from Table 2 in Rajeg and Rajeg (2019:71)
a <- 9
b <- 1
c <- 408
d <- 305

logOR_compute(a, b, c, d)

# the above data produces positive log OR
# Observed freq. for cell "a" is larger than the Expected freq.

# Data from Table 3 in Rajeg and Rajeg (2019:74)
a <- 1
b <- 6
c <- 416
d <- 300

logOR_compute(a, b, c, d)

# the above data produces negative log OR
# Expected freq. for cell "a" is larger than the Observed freq.

}
\references{
\itemize{
      \item Rajeg, Gede Primahadi Wijaya & I Made Rajeg. 2019. Analisis Koleksem Khas dan potensinya untuk kajian kemiripan makna konstruksional dalam Bahasa Indonesia. In I Nengah Sudipa (ed.), \emph{ETIKA BAHASA Buku persembahan menapaki usia pensiun: I Ketut Tika}, vol. 1, 65–83. Denpasar, Bali, Indonesia: Swasta Nulus. \url{https://doi.org10.26180/5bf4e49ea1582}. \url{https://osf.io/preprints/inarxiv/uwzts/} (30 January, 2019).
      \item Gries, S. T. (2014). Coll.analysis 3.5. A script for R to compute perform collostructional analyses. \url{http://www.linguistics.ucsb.edu/faculty/stgries/teaching/groningen/index.html}.
}
}
