% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collex_MI.R
\name{MI_compute}
\alias{MI_compute}
\title{Compute Mutual Information}
\usage{
MI_compute(a, a_exp)
}
\arguments{
\item{a}{observed frequency co-occurrence of a collocate and the nodeword.}

\item{a_exp}{expected frequency co-occurrence of the collocate and the nodeword.}
}
\value{
A double-numeric value of MI score
}
\description{
An internal function called by \code{\link{collex_MI}} to compute Mutual Information (MI) score
}
\examples{
# the numbers below are made-up examples
# for the collocate pattern "tindak pidana" 'legal act'
a <- 702
total_freq_w1 <- 1183
total_freq_w2 <- 1251
corpus_size <- 5763049
a_exp <- (total_freq_w1 * total_freq_w2)/corpus_size
(MI_score <- MI_compute(a, a_exp))
}
