% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{node_vect_pos}
\alias{node_vect_pos}
\title{Gather the node-vector}
\usage{
node_vect_pos(df = NULL, regex_pattern = NULL)
}
\arguments{
\item{df}{The tibble from the output of \code{word_vect_to_df}.}

\item{regex_pattern}{Passed on from the regular expression specified into \code{pattern} argument of \code{colloc_leipzig}.}
}
\value{
An integer vector.
}
\description{
Utility function to determine the vector-position of the search pattern/node word for which the collocates to be extracted.
    It is called internally by \code{colloc_leipzig}
}
