% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{placeholder}
\alias{placeholder}
\title{Output placeholder}
\usage{
placeholder(input = NULL, input_names = NULL, mode = "list")
}
\arguments{
\item{input}{The input object verified via another utility function, namely \code{check_file_input}.}

\item{input_names}{Names of the corpus file}

\item{mode}{Type of object for the placeholder. By default, it is \code{"list"}.}
}
\value{
empty list for the results of \code{colloc_leipzig}.
}
\description{
Utility function to generate the placeholder/object for the output elements of \code{colloc_leipzig}.
    It is called internally by \code{colloc_leipzig}.
}
