% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_output_file}
\alias{save_output_file}
\title{Saving outputs}
\usage{
save_output_file(raw_colloc = NULL, freq_colloc = NULL, freqlist = NULL,
  corpussize = NULL, sent_match = NULL)
}
\arguments{
\item{raw_colloc}{File name for the raw collocates table.
It is passed on from \code{colloc_output_name} argument of \code{colloc_leipzig}.}

\item{freq_colloc}{File name for the collocates frequency table.
It is passed on from \code{colloc_freq_output_name} argument of \code{colloc_leipzig}.}

\item{freqlist}{File name for the frequency list table of all word tokens in the corpus.
It is passed on from \code{freqlist_output_name} argument of \code{colloc_leipzig}.}

\item{corpussize}{File name for the overall corpus size in word-tokens.
It is passed on from \code{corpus_size_output_name} argument of \code{colloc_leipzig}.}

\item{sent_match}{File name for the sentence match in which the \code{pattern} argument is detected.
It is passed on from \code{sentence_output_name} argument of \code{colloc_leipzig}.}
}
\value{
Tab-separated plain texts corresponding to the specified output results.
}
\description{
Utility function to generate the output file of \code{colloc_leipzig} when the \code{save_results} argument is set to \code{TRUE}.
    It is called internally by \code{colloc_leipzig}.
}
