% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collex_TScore.R
\name{Tscore_compute}
\alias{Tscore_compute}
\title{Compute T-Score}
\usage{
Tscore_compute(a, b, c, d)
}
\arguments{
\item{a}{frequency co-occurrence of a collocate and the nodeword.}

\item{b}{overall frequency of the collocate in the corpus}

\item{c}{overall frequency of the nodeword in the corpus}

\item{d}{total size of the corpus}
}
\value{
A double-numeric value of T-Score
}
\description{
An internal function called by \code{\link{collex_TScore}} to compute T-Score values
}
\examples{
# the numbers below are made-up examples
# for the collocate pattern "tindak pidana" 'legal act'
a <- 702
b <- 1183
c <- 1251
d <- 5763049
(tscore <- Tscore_compute(a, b, c, d))
}
