% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{word_vect_to_df}
\alias{word_vect_to_df}
\title{Data frame function for word-vector}
\usage{
word_vect_to_df(tokenised_text = NULL)
}
\arguments{
\item{tokenised_text}{The output of \code{tokenising}.}
}
\value{
A tbl_df.
}
\description{
Utility function to turn the tokenised corpus into data frame of word-tokens.
    It is called internally by \code{colloc_leipzig}.
}
