% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fye.R
\name{fye_compute}
\alias{fye_compute}
\title{Fisher's Exact Test}
\usage{
fye_compute(a, a_exp, b, c, d)
}
\arguments{
\item{a}{observed co-occurrence frequency of a given collocate and the nodeword.}

\item{a_exp}{expected co-occurrence frequency of the collocate and the nodeword.}

\item{b}{co-occurrence frequency of the collocate with other words in the corpus.}

\item{c}{co-occurrence frequency of the nodeword with other words in the corpus.}

\item{d}{co-occurrence frequency of other words in the corpus.}
}
\value{
Numeric vectors of p-values
}
\description{
Perform one-tailed Fisher's Exact test for the collostruction/collocation strength.
    The \emph{p}-value is log-transformed to the base of ten as in the \emph{Collostructional Analysis}.
    This is an internal function called via \code{\link{collex_fye}}.
}
