% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fye_compute}
\alias{fye_compute}
\title{Fisher's Exact Test}
\usage{
fye_compute(df, mpfr_precision = NULL, collstr_digit = NULL)
}
\arguments{
\item{df}{The output of \code{\link{assoc_prepare}}.}

\item{mpfr_precision}{Integer indicating the maximal precision to be used in \bold{bits}. This is passed to the \code{precBits} argument of \code{\link[Rmpfr]{mpfr}}.}

\item{collstr_digit}{The floating digits of the collostruction strength. It is passed on from \code{\link{collex_fye}} and the default is \code{3}.}
}
\value{
A double vector of collostruction strength
}
\description{
Perform one-tailed Fisher's Exact test for the collostruction/collocation strength.
    The \emph{p}-value is log-transformed to the base of ten as in the \emph{Collostructional Analysis}.
    This is an internal function called via \code{\link{collex_fye}}.
}
