% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{leipzig_corpus_path}
\alias{leipzig_corpus_path}
\title{Filepath of the Indonesian Leipzig Corpora}
\format{An object of class \code{character} of length 15.}
\usage{
leipzig_corpus_path
}
\description{
The author's full filepaths to the plain-texts of the Indonesian Leipzig Corpora
}
\keyword{datasets}
