% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collex_MI.R
\name{MI_compute}
\alias{MI_compute}
\title{Compute Mutual Information}
\usage{
MI_compute(a, b, c, d)
}
\arguments{
\item{a}{frequency co-occurrence of a collocate and the nodeword.}

\item{b}{overall frequency of the collocate in the corpus}

\item{c}{overall frequency of the nodeword in the corpus}

\item{d}{total size of the corpus}
}
\value{
A double-numeric value of MI score
}
\description{
An internal function called by \code{\link{collex_MI}} to compute Mutual Information (MI) score
}
\examples{
# the numbers below are made-up examples
# for the collocate pattern "tindak pidana" 'legal act'
a <- 702
b <- 1183
c <- 1251
d <- 5763049
(MI_score <- MI_compute(a, b, c, d))
}
