% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Filepath of the cleaned Indonesian Leipzig Corpora}
\format{An object of class \code{character} of length 398.}
\usage{
stopwords
}
\description{
Package author's filepath to the cleaned version of the Indonesian Leipzig Corpora. The cleaned version means that the non-word tokens in the corpus (e.g. punctuation) have been separated by white-space.
Filepath of the frequency list of the Indonesian Leipzig Corpora
}
\details{
Package author's filepath to the frequency list files of the Indonesian Leipzig Corpora. These files are generated by the author using \emph{MonARCH High Performance Computing Cluster} at Monash University, Australia.
Filepath of the corpus-word vector the Indonesian Leipzig Corpora

Package author's filepath to the corpus-word vector data of the Indonesian Leipzig Corpora. These files are the tokenised versions of the Indonesian Leipzig Corpora and are generated by the author using \emph{MonARCH High Performance Computing Cluster} at Monash University, Australia.
Filepath of the Balinese newspapers corpus

Package author's filepath to the Balinese newspapers texts.
Indonesian stopwords

Character vector containing Indonesian words selected as stopwords.
}
\keyword{datasets}
