% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stopwords}
\alias{stopwords}
\title{Indonesian stopwords}
\format{An object of class \code{character} of length 398.}
\usage{
stopwords
}
\description{
Character vector containing Indonesian words selected as stopwords.
}
\keyword{datasets}
