% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collex_TScore.R
\name{Tscore_compute}
\alias{Tscore_compute}
\title{Compute T-Score}
\usage{
Tscore_compute(a, a_exp)
}
\arguments{
\item{a}{frequency co-occurrence of a collocate and the nodeword.}

\item{a_exp}{expected frequency co-occurrence of the collocate and the nodeword.}
}
\value{
A double-numeric value of T-Score
}
\description{
An internal function called by \code{\link{collex_TScore}} to compute T-Score values
}
\examples{
# the numbers below are made-up examples
# for the collocate pattern "tindak pidana" 'legal act'
a <- 702
total_freq_w1 <- 1183
total_freq_w2 <- 1251
corpus_size <- 5763049
a_exp <- (total_freq_w1 * total_freq_w2)/corpus_size
(tscore <- Tscore_compute(a, a_exp))
}
